refactor: token definitions moved to its own file
