+++ src/lexer/lexer.ts
@@ -12,118 +12,7 @@ import { isAlpha, isNumeric, isAlphaNum, isSpace, isHexadecimal, inArray } from 
 import LexingError from "@lexer/lexer-error.ts"
 import LexerCursor from "@lexer/lexer-cursor.ts"
 
-export type TokenType = |
-  "IdentifierName"      |
-  "Statement"           |
-  "NumericLiteral"      |
-  "FloatLiteral"        |
-  "BinaryLiteral"       |
-  "OctalLiteral"        |
-  "HexaLiteral"         |
-  "CharLiteral"         |
-  "StringLiteral"       |
-  "TemplateString"      |
-  "Keyword"             |
-  "Operator"            |
-  "Token"
-
-export enum Tokens {
-  LPAREN        = '(',
-  RPAREN        = ')',
-  LBRACE        = '{',
-  RBRACE        = '}',
-  LSQRBR        = '[',
-  RSQRBR        = ']',
-  COMMA         = ',',
-  DQUOTE        = '"',
-  QUOTE         = '\'',
-  BACKSTICK     = '`',
-  PLUS          = '+',
-  MINUS         = '-',
-  STAR          = '*',
-  SLASH         = '/',
-  PERCEN        = '%',
-  LESS          = '<',
-  GREATER       = '>',
-  DOT           = '.',
-  AMPER         = '&',
-  EQUAL         = '=',
-  ESP           = '~',
-  COLON         = ':',
-  SEMICOLON     = ';',
-  EXCLAMATION   = '!',
-  INTERROGATION = '?',
-  UNDERSCORE    = '_',
-
-  // Spacials Tokens (e.g: module accessor)
-  MODULE_ACCESSOR = "::",
-  INCREMENT       = "++",
-  DECREMENT       = "--",
-
-  ADD_ASSIGN      = "+=",
-  MINUS_ASSIGN    = "-=",
-  MULT_ASSIGN     = "*=",
-  DIV_ASSIGN      = "/="
-}
-
-export enum Keywords {
-  FN          = "fn",
-  VAR         = "var",
-  CONST       = "const",
-  RETURN      = "return",
-  IMPORT      = "import",
-  FROM        = "from",
-  EXPORT      = "export",
-  CLASS       = "class",
-  THIS        = "this",
-  SUPER       = "super",
-  TYPE        = "type",
-  INTERFACE   = "interface",
-  ENUM        = "enum",
-  EXTENDS     = "extends",
-  IMPLEMENTS  = "implements",
-  MODULE      = "module",
-  FOR         = "for",
-  WHILE       = "while",
-  DO          = "do",
-  BREAK       = "break",
-  SKIP        = "skip",
-  IF          = "if",
-  ELSE        = "else",
-  ELIF        = "elif",
-  MATCH       = "match",
-  SWITCH      = "switch",
-  CASE        = "case",
-  DEFAULT     = "default",
-  IN          = "in",
-  OF          = "of",
-  AND         = "and",
-  OR          = "or",
-  NOT         = "not",
-  TRY         = "try",
-  CATCH       = "catch",
-  AS          = "as",
-  CONSTRUCTOR = "constructor",
-  DESTRUCTOR  = "destructor",
-  NEW         = "new",
-  DROP        = "drop",
-  PUBLIC      = "public",
-  PRIVATE     = "private",
-  PROTECTED   = "protected",
-  STATIC      = "static",
-  OVERRIDE    = "override",
-  SETTER      = "setter",
-  GETTER      = "getter",
-  ASYNC       = "async",
-  AWAIT       = "await"
-}
-
-export interface Token {
-  type: TokenType,
-  content: string,
-  pos: number,
-  line: number
-}
+import { isAlpha, isNumeric, isAlphaNum, isSpace, isHexadecimal, inArray } from "@utils"
 
 const VALID_HEXADECIMAL_END = [
   'A', 'B', 'C', 'D', 'E', 'F'
